requirements.txt
# pip==21.2.4
# setuptools==65.5.0
numpy==1.21.4
numba==0.53.0
tqdm
fire==0.4.0
gin-config==0.4.0
logdir==0.13.0
ribs[all]==0.5.0
gym==0.18.0
Pillow==7.2.0
pandas==1.5.3
networkx==3.1
cma==3.3.0
gurobipy==11.0.0

# Pytorch
-f https://download.pytorch.org/whl/cu117/torch_stable.html
torch==1.13.0+cu117
torchvision==0.14.0+cu117
torchaudio==0.13.0+cu117

# Dask
dask==2.30.0
dask-jobqueue==0.7.1
distributed==2.30.0
click==7.1.2  # Newer click causes error with dask scheduler.
bokeh==2.2.3
jupyter-server-proxy==1.5.0

# Plot
matplotlib==3.4.2
seaborn==0.12.0
loguru==0.5.3
pingouin==0.5.4
opencv-python==4.7.0.72
imageio==2.9.0
moviepy==1.0.3

gym-minigrid==1.0.1
stable-baselines3==2.2.1
scikit-image==0.19.2

wandb==0.17.0